fix: as suggested switched to using StringSlice and cache_shared.go
